+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/testing/CarJavaInterfaceChecker.java
@@ -1,0 +1,83 @@
+package com.mapbox.navigation.qa_test_app.testing;
+
+import androidx.annotation.NonNull;
+import androidx.car.app.CarContext;
+import androidx.lifecycle.LifecycleOwner;
+
+import com.mapbox.androidauto.MapboxCarNavigationManager;
+import com.mapbox.androidauto.internal.car.search.CarPlaceSearch;
+import com.mapbox.maps.extension.androidauto.MapboxCarMap;
+import com.mapbox.navigation.base.options.NavigationOptions;
+import com.mapbox.navigation.core.MapboxNavigation;
+import com.mapbox.navigation.core.lifecycle.MapboxNavigationApp;
+import com.mapbox.navigation.core.lifecycle.MapboxNavigationObserver;
+
+import java.util.List;
+
+import kotlinx.coroutines.CoroutineScope;
+
+class CarJavaInterfaceChecker {
+
+  void MapboxNavigationApp(
+          LifecycleOwner lifecycleOwner,
+          NavigationOptions navigationOptions,
+          MapboxNavigationObserver observer
+  ) {
+    // Set up now
+    MapboxNavigationApp.setup(navigationOptions);
+
+    // Set up provider
+    MapboxNavigationApp.setup(() -> navigationOptions);
+
+    // Control lifecycles
+    MapboxNavigationApp.attach(lifecycleOwner);
+    MapboxNavigationApp.disable();
+    MapboxNavigationApp.detach(lifecycleOwner);
+
+    // Get current instance
+    MapboxNavigation mapboxNavigation = MapboxNavigationApp.current();
+
+    // Register and unregister observer
+    MapboxNavigationApp.registerObserver(observer);
+    MapboxNavigationApp.unregisterObserver(observer);
+
+    MapboxNavigationObserver otherObserver = MapboxNavigationApp.getObserver(CarPlaceSearch.class);
+    List<CarPlaceSearch> otherObservers = MapboxNavigationApp.getObservers(CarPlaceSearch.class);
+  }
+
+  void MapboxNavigationObserver() {
+    MapboxNavigationObserver observer = new MapboxNavigationObserver() {
+      @Override
+      public void onAttached(@NonNull MapboxNavigation mapboxNavigation) {
+
+      }
+
+      @Override
+      public void onDetached(@NonNull MapboxNavigation mapboxNavigation) {
+
+      }
+    };
+  }
+
+  void MapboxCarNavigationManager(
+          CarContext carContext,
+          LifecycleOwner lifecycleOwner,
+          MapboxCarMap mapboxCarMap
+  ) {
+    // Constructor
+    MapboxCarNavigationManager sut = new MapboxCarNavigationManager(carContext);
+
+    // Observing auto drive
+    CoroutineScope scope = JavaFlow.lifecycleScope(lifecycleOwner);
+    JavaFlow.collect(sut.getAutoDriveEnabledFlow(), scope, (enabled) -> {
+      // check enabled
+    });
+
+    // Get auto drive value
+    boolean isEnabled = sut.getAutoDriveEnabledFlow().getValue();
+
+    // Register onto MapboxNavigationAPp
+    MapboxNavigationApp.registerObserver(sut);
+    MapboxNavigationApp.unregisterObserver(sut);
+  }
+}